dropjob: upstream ark behaviour from various file view implementations
For whatever reason, a lot of file views in KDE specialcase dropping Ark's
MIME data but haven't put that duplicated logic upstream here in the dropjob
where it belongs. This rectifies that, and new file views that rely on KIO::drop
for drag & drop functionality no longer have to duplicate this code.